v0.0.03 support for curve application in vcol
this is only available for blender r60055 or later (the script checks if
this is so). The reason for this is that it previous versions it might
happen that an uninitialized curvemapping was accessed leading to a hard
crash. Never revisions support an initialize function (and won't crash
but throw an exception)