ensure that twinning respects width
Currently twinning only correctly positions gears in z if the width of the driver is 0.1 yet the default is 0.2. The default of 0.2 means that the twinned gears overlap in the z which means further gears in the chain tend to end up intersecting. The proposed change takes into account the width of the current gear and the driver when calculating the z offset. This correctly positions this gear in respect to the driver regardless of the width of either. What it doesn't solve, however, is that further gears in the train with differing widths may intersect. Ideally a "proper" fix would work all the way backwards to the root gear to ensure that not only is each gear positioned correctly against the driver but so that it also doesn't intersect with any previous gears.